% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{byrow}
\alias{byrow}
\alias{byrow-package}
\title{byrow: Compute various statistics by row.}
\description{
This package is inspired by the \code{rowMeans()} function, and provides
functions to compute other statistics by row, such as the standard deviation.
}
