% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowSds.R
\name{rowSds}
\alias{rowSds}
\title{Compute the standard deviations of the rows of a data frame.}
\usage{
rowSds(l)
}
\arguments{
\item{l}{A numeric data frame (or a list with equally long elements)}
}
\value{
A list with the row-wise standard deviations.
}
\description{
Compute the standard deviations of the rows of a data frame.
}
