% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowZscs.R
\name{rowZscs}
\alias{rowZscs}
\title{Compute row-wise Z-scores and p-valuse using Stouffer's method.}
\usage{
rowZscs(l)
}
\arguments{
\item{l}{A numeric data frame of p-values (or a list with equally long
elements)}
}
\value{
A data frame with two columns: rho_zscore and pred_perf_pval
}
\description{
Compute row-wise Z-scores and p-valuse using Stouffer's method.
}
